working stage 1 - separated timer code
