+++ index.html
@@ -22,13 +22,15 @@
             <button onclick="startTimer()">Start Timer</button>
             <button onclick="stopTimer()">Stop Timer</button>
         </div> -->
-        <section id="timer">
+        <!-- <section id="timer">
           <div id="timer-display">00:00</div>
           <button id="start-timer">Start Timer</button>
           <button id="stop-timer" disabled>Stop Timer</button>
-        </section>
+        </section> -->
+        <div id="timer-control-container"></div>
     </div>
     <script type="module" src="scripts/db.js"></script>
+    <script type="module" src="scripts/timerControl.js"></script>
     <script type="module" src="scripts/app.js"></script>
 </body>
 </html>
