+++ timerControl.html
@@ -1,0 +1,6 @@
+<!-- timerControl.html -->
+<div id="timer-control" class="timer-control">
+  <div id="timer-display">00:00</div>
+  <button id="start-timer">Start Timer</button>
+  <button id="stop-timer" disabled>Stop Timer</button>
+</div>
